Add project tags
