+++ Assets/styles/cursor.scss
@@ -1,0 +1,16 @@
+.cursor {
+  pointer-events: none;
+  z-index: 999999;
+
+  &__ball {
+    position: fixed;
+    top: 0;
+    left: 0;
+    mix-blend-mode: difference;
+    z-index: 999999;
+
+    circle {
+      fill: #f7f8fa;
+    }
+  }
+}
