+++ Assets/styles/index.scss
@@ -1,0 +1,2 @@
+@import "./main.scss";
+@import "./cursor.scss";
