+++ Assets/styles/main.scss
@@ -38,11 +38,13 @@ input {
   margin: 0;
   padding: 0;
   list-style-type: none;
-  /* outline: 1px solid green; */
+  text-decoration: none;
+  cursor: initial;
+  outline: 1px solid green;
 }
 a {
   color: var(--color-dark);
-  text-decoration: none;
+  cursor: pointer;
 }
 dialog {
   outline: none;
@@ -82,6 +84,7 @@ html {
 }
 body {
   width: 100vw;
+  user-select: none;
 }
 .wrapper {
   width: 100vw;
@@ -182,7 +185,7 @@ body {
   height: 100%;
   opacity: 0.86;
   padding: unset;
-  cursor: pointer;
+  // cursor: pointer;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -570,6 +573,7 @@ body {
   margin: 0;
   font-size: 23px;
   font-weight: medium;
+  text-transform: uppercase;
 }
 .item__subtext {
   opacity: 0.56;
@@ -584,6 +588,26 @@ body {
   font-style: 12px;
   font-weight: book;
   opacity: 60%;
+}
+.tags {
+  position: relative;
+  display: grid;
+  grid-auto-flow: column;
+  position: absolute;
+  z-index: 100;
+  top: 2rem;
+  left: 2rem;
+  gap: 0.5rem;
+  &__head {
+    position: absolute;
+    top: -2rem;
+    opacity: 0;
+  }
+  &__item {
+    padding: 0.5rem;
+    background-color: #fff;
+    border-radius: 4px;
+  }
 }
 .links {
   display: flex;
