+++ index.html
@@ -127,10 +127,9 @@
     <link rel="manifest" href="/manifest.json" />
     <title>Simo Ngquseka</title>
     <link rel="stylesheet" prefetch href="Assets/fonts/fonts.css" />
-    <link rel="stylesheet" type="text/css" href="Assets/styles/main.css" />
   </head>
 
-  <body class="loading">
+  <body>
     <!-- <div class="wrapper"> -->
     <div class="loader flex">
       <h3 class="centered pulse">LOADING</h3>
@@ -207,6 +206,14 @@
         <div class="sect-main">
           <div class="card preview">
             <!-- <canvas id="proj-preview" class="img img--preview preview--lg"></canvas> -->
+            <div class="tags">
+              <div class="tags__head">Used</div>
+              <span class="tags__item"></span>
+              <span class="tags__item"></span>
+              <span class="tags__item"></span>
+              <span class="tags__item"></span>
+            </div>
+
             <img
               src="Assets/images/loader.gif"
               id="proj-preview"
